requirements.txt
accelerate
bitsandbytes
datasets
dotenv
peft
timm
trl==0.14.0
transformers
-e .
python-box
matplotlib
seaborn

unsloth
torch
xformers
triton
cut-cross-entropy
sentencepiece
protobuf
huggingface_hub>=0.34.0
hf_transfer
pyyaml
optuna
wandb

